fetchai_agents/requirements.txt
# ABOUTME: Python dependencies for Fetch.ai agent deployment.
# ABOUTME: Install with: pip install -r requirements.txt

# Fetch.ai uAgents framework
uagents>=0.12.0

# HTTP client for backend API calls
httpx>=0.27.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0

# HTTP client for registration
requests>=2.31.0
